front and back end most of the way. Went for making a duplicate of the current instance values and settings, and storing against UUID in the database. There are various options from here I thihnk
for examlpe, just erturning the completed SVG (this is only possible when saved from the front end, not currently if the thing is built by the API (which doesnt exist :) ) , or returning an iframe which animates (on mouseover? ).

also it seems like a good idea to save the current toggle features and title string actually into the DB model instance. Ill do that tomorrow aswell.

should be live tomorrow I think